P1: extract cos/tx_completion.rs (17 fns + 2 consts, ~430 LOC)
Closes #956 Phase 6 builder edge (cos/builders.rs -> tx::cos_tick_for_ns)
and the TX-completion / timer-wheel subset of Phase 7 deferrals
(10 of 18 fns imported by cos/queue_service.rs from crate::afxdp::tx::
move to super::tx_completion::).

Move set:
- Timer wheel: cos_tick_for_ns, cos_timer_wheel_level_and_slot,
  wake_cos_queue, count_tx_ring_full_submit_stall, rearm_cos_queue,
  mark_cos_queue_runnable, normalize_cos_queue_state,
  advance_cos_timer_wheel, cascade_cos_timer_wheel_level1,
  wake_due_cos_timer_slot + COS_TIMER_WHEEL_TICK_NS +
  COS_TIMER_WHEEL_L0_HORIZON_TICKS.
- TX-completion: prime_cos_root_for_service,
  apply_direct_exact_send_result, refresh_cos_interface_activity,
  apply_cos_send_result, apply_cos_prepared_result,
  restore_cos_local_items_inner, restore_cos_prepared_items_inner.

Visibility: 14 items pub(in crate::afxdp), 4 fns + 1 const file-private.
#[inline] adds on hot-path / per-batch fns per Phase 4-8 lesson; preserves
existing #[inline] on count_tx_ring_full_submit_stall.

cargo build --bins clean (no new unused-import warnings).
cargo test --bins 865/0/2 (rolling baseline).

Plan: docs/pr/p1-tx-completion/plan.md (v5.3, triadic-converged).
