+++ userspace-dp/src/afxdp/cos/queue_service.rs
@@ -20,13 +20,17 @@
 // LLVM's heuristic threshold should cover them; revisit only if a
 // post-merge perf regression points at one.
 //
-// Back-edges to tx.rs are deferred TX-completion + worker-binding
-// helpers (apply_cos_*_result, restore_cos_*_inner,
-// prime_cos_root_for_service, advance_cos_timer_wheel, transmit_*,
-// reap_tx_completions, refresh_cos_interface_activity, etc).
-// Visibility on those bumped to pub(in crate::afxdp) in tx.rs as
-// part of this PR. They will move with the future TX-completion
-// extraction.
+// TX-completion + timer-wheel back-edges (apply_cos_*_result,
+// restore_cos_*_inner, prime_cos_root_for_service,
+// refresh_cos_interface_activity, cos_tick_for_ns +
+// cos_timer_wheel_level_and_slot + count_tx_ring_full_submit_stall)
+// moved to cos/tx_completion.rs in #956 P1.
+//
+// Remaining back-edges to crate::afxdp::tx are XSK-ring /
+// worker-binding / prepared-frame primitives (transmit_*,
+// reap_tx_completions, maybe_wake_tx, recycle_*, stamp_submits,
+// cos_queue_dscp_rewrite, TxError, the guarantee/quantum constants).
+// Those move with the afxdp/tx/ split in #984.
 
 use std::collections::VecDeque;
 use std::sync::atomic::Ordering;
@@ -53,17 +57,23 @@ use super::{
     refill_cos_tokens, COS_MIN_BURST_BYTES,
 };
 
-// Back-edges to tx.rs (Phase 7 deferrals — TX-completion + worker-
-// binding family). Visibility bumped to pub(in crate::afxdp) so this
-// module can reach them.
-use crate::afxdp::tx::{
+// #956 P1: TX-completion + timer-wheel symbols moved to
+// cos/tx_completion.rs.
+use super::tx_completion::{
     apply_cos_prepared_result, apply_cos_send_result,
-    apply_direct_exact_send_result, cos_queue_dscp_rewrite, cos_tick_for_ns,
-    cos_timer_wheel_level_and_slot, count_tx_ring_full_submit_stall, maybe_wake_tx,
-    prime_cos_root_for_service, reap_tx_completions, recycle_cancelled_prepared_offset,
-    refresh_cos_interface_activity, remember_prepared_recycle, restore_cos_local_items_inner,
-    restore_cos_prepared_items_inner, stamp_submits, transmit_batch, transmit_prepared_queue,
-    TxError, COS_GUARANTEE_QUANTUM_MAX_BYTES, COS_GUARANTEE_QUANTUM_MIN_BYTES,
+    apply_direct_exact_send_result, cos_tick_for_ns,
+    cos_timer_wheel_level_and_slot, count_tx_ring_full_submit_stall,
+    prime_cos_root_for_service, refresh_cos_interface_activity,
+    restore_cos_local_items_inner, restore_cos_prepared_items_inner,
+};
+// Remaining back-edges to tx.rs (XSK-ring / worker-binding /
+// prepared-frame primitives + TxError + guarantee/quantum constants —
+// deferred to #984 / afxdp/tx/ split).
+use crate::afxdp::tx::{
+    cos_queue_dscp_rewrite, maybe_wake_tx, reap_tx_completions,
+    recycle_cancelled_prepared_offset, remember_prepared_recycle, stamp_submits,
+    transmit_batch, transmit_prepared_queue, TxError,
+    COS_GUARANTEE_QUANTUM_MAX_BYTES, COS_GUARANTEE_QUANTUM_MIN_BYTES,
     COS_GUARANTEE_VISIT_NS, COS_SURPLUS_ROUND_QUANTUM_BYTES,
 };
