+++ userspace-dp/src/afxdp/tx.rs
@@ -1124,87 +1124,12 @@ fn ingest_cos_pending_tx_with_provenance(
     bound_pending_tx_local(binding);
 }
 
-pub(in crate::afxdp) fn prime_cos_root_for_service(binding: &mut BindingWorker, root_ifindex: i32, now_ns: u64) -> bool {
-    let shared_root_lease = binding
-        .cos_fast_interfaces
-        .get(&root_ifindex)
-        .and_then(|iface_fast| iface_fast.shared_root_lease.clone());
-    let Some(root) = binding.cos_interfaces.get_mut(&root_ifindex) else {
-        return false;
-    };
-    advance_cos_timer_wheel(root, now_ns);
-    if let Some(shared_root_lease) = shared_root_lease.as_ref() {
-        maybe_top_up_cos_root_lease(root, shared_root_lease, now_ns);
-    }
-    true
-}
 
-pub(in crate::afxdp) fn apply_direct_exact_send_result(
-    binding: &mut BindingWorker,
-    root_ifindex: i32,
-    queue_idx: usize,
-    sent_packets: u64,
-    sent_bytes: u64,
-) {
-    if let Some(root) = binding.cos_interfaces.get_mut(&root_ifindex) {
-        if let Some(queue) = root.queues.get_mut(queue_idx) {
-            queue.queued_bytes = queue.queued_bytes.saturating_sub(sent_bytes);
-            queue.tokens = queue.tokens.saturating_sub(sent_bytes);
-            // #760 instrumentation: record the exact-owner-local
-            // send at the same place the token bucket decrements.
-            // Divide by a scrape window to get an observed per-queue
-            // drain rate and compare against
-            // `queue.transmit_rate_bytes` to detect a cap bypass.
-            queue
-                .owner_profile
-                .drain_sent_bytes
-                .fetch_add(sent_bytes, Ordering::Relaxed);
-        }
-        root.tokens = root.tokens.saturating_sub(sent_bytes);
-    }
-    if let Some(shared_root_lease) = binding
-        .cos_fast_interfaces
-        .get(&root_ifindex)
-        .and_then(|iface_fast| iface_fast.shared_root_lease.as_ref())
-    {
-        shared_root_lease.consume(sent_bytes);
-    }
-    if let Some(shared_queue_lease) = binding
-        .cos_fast_interfaces
-        .get(&root_ifindex)
-        .and_then(|iface_fast| iface_fast.queue_fast_path.get(queue_idx))
-        .and_then(|queue_fast| queue_fast.shared_queue_lease.as_ref())
-    {
-        shared_queue_lease.consume(sent_bytes);
-    }
-    refresh_cos_interface_activity(binding, root_ifindex);
-    if sent_packets > 0 {
-        binding
-            .live
-            .tx_packets
-            .fetch_add(sent_packets, Ordering::Relaxed);
-        binding
-            .live
-            .tx_bytes
-            .fetch_add(sent_bytes, Ordering::Relaxed);
-        // #760 instrumentation, exact-owner-local path. Paired with
-        // tx_bytes unconditionally — if the per-queue drain_sent_bytes
-        // above (guarded by `if let Some(queue)`) ever undercounts
-        // this, the gap is an `apply_*` early-return / queue-miss.
-        binding
-            .live
-            .owner_profile_owner
-            .drain_sent_bytes_shaped_unconditional
-            .fetch_add(sent_bytes, Ordering::Relaxed);
-    }
-}
 
-const COS_TIMER_WHEEL_TICK_NS: u64 = 50_000;
 pub(in crate::afxdp) const COS_GUARANTEE_VISIT_NS: u64 = 200_000;
 pub(in crate::afxdp) const COS_GUARANTEE_QUANTUM_MIN_BYTES: u64 = 1500;
 pub(in crate::afxdp) const COS_GUARANTEE_QUANTUM_MAX_BYTES: u64 = 512 * 1024;
 pub(in crate::afxdp) const COS_SURPLUS_ROUND_QUANTUM_BYTES: u64 = 1500;
-const COS_TIMER_WHEEL_L0_HORIZON_TICKS: u64 = COS_TIMER_WHEEL_L0_SLOTS as u64;
 
 // #956: cos/ submodule imports.
 //
@@ -1227,12 +1152,11 @@ const COS_TIMER_WHEEL_L0_HORIZON_TICKS: u64 = COS_TIMER_WHEEL_L0_SLOTS as u64;
 use super::cos::{
     apply_cos_admission_ecn_policy, cos_flow_aware_buffer_limit, cos_flow_bucket_index,
     cos_item_flow_key, cos_queue_drain_all, cos_queue_flow_share_limit,
-    cos_queue_is_empty, cos_queue_push_back, cos_queue_push_front, cos_queue_restore_front,
-    drain_shaped_tx, ensure_cos_interface_runtime, maybe_top_up_cos_root_lease, park_cos_queue,
-    prepared_cos_request_stays_on_current_tx_binding, publish_committed_queue_vtime,
-    redirect_local_cos_request_to_owner, redirect_prepared_cos_request_to_owner,
-    redirect_prepared_cos_request_to_owner_binding, release_cos_root_lease,
-    resolve_local_routing_decision, CoSServicePhase, LocalRoutingDecision, Step1Action,
+    cos_queue_is_empty, cos_queue_push_back, cos_queue_restore_front,
+    drain_shaped_tx, ensure_cos_interface_runtime, mark_cos_queue_runnable,
+    publish_committed_queue_vtime, redirect_prepared_cos_request_to_owner,
+    redirect_prepared_cos_request_to_owner_binding,
+    resolve_local_routing_decision, LocalRoutingDecision, Step1Action,
 };
 #[cfg(test)]
 use super::cos::ecn::{ethernet_l3, mark_ecn_ce_ipv4, mark_ecn_ce_ipv6, EthernetL3};
@@ -1257,171 +1181,18 @@ use super::cos::{
     COS_FLOW_FAIR_MIN_SHARE_BYTES, ECN_CE, ECN_ECT_0, ECN_ECT_1, ECN_MASK, ECN_NOT_ECT,
     V_MIN_CONSECUTIVE_SKIP_HARD_CAP, V_MIN_SUSPENSION_BATCHES,
 };
-
-// #956 Phase 6: visibility bumped from private so cos/builders.rs
-// can reach it via `use crate::afxdp::tx::cos_tick_for_ns`. Timer-
-// wheel helpers stay in tx.rs through the drain-scheduler phase
-// (they consume the private COS_TIMER_WHEEL_* constants and are
-// tightly coupled to advance_cos_timer_wheel).
-pub(in crate::afxdp) fn cos_tick_for_ns(now_ns: u64) -> u64 {
-    now_ns / COS_TIMER_WHEEL_TICK_NS
-}
-
-pub(in crate::afxdp) fn cos_timer_wheel_level_and_slot(current_tick: u64, wake_tick: u64) -> (u8, usize) {
-    if wake_tick.saturating_sub(current_tick) < COS_TIMER_WHEEL_L0_HORIZON_TICKS {
-        (0, (wake_tick % COS_TIMER_WHEEL_L0_SLOTS as u64) as usize)
-    } else {
-        (
-            1,
-            ((wake_tick / COS_TIMER_WHEEL_L0_SLOTS as u64) % COS_TIMER_WHEEL_L1_SLOTS as u64)
-                as usize,
-        )
-    }
-}
-
-// #956 Phase 2: flow-hashing helpers (mix_cos_flow_bucket,
-// exact_cos_flow_bucket, cos_flow_hash_seed_from_os,
-// cos_item_flow_key, cos_flow_bucket_index,
-// cos_queue_prospective_active_flows) extracted to
-// userspace-dp/src/afxdp/cos/flow_hash.rs. Production callers
-// import via super::cos::{...}; tests import the same way.
-//
-// Constants (COS_FLOW_FAIR_BUCKETS / COS_FLOW_FAIR_BUCKET_MASK)
-// stayed in afxdp::types — flow_hash.rs imports them.
-
-fn wake_cos_queue(root: &mut CoSInterfaceRuntime, queue_idx: usize) {
-    let Some(queue) = root.queues.get_mut(queue_idx) else {
-        return;
-    };
-    if cos_queue_is_empty(queue) {
-        queue.runnable = false;
-        queue.parked = false;
-        queue.next_wakeup_tick = 0;
-        return;
-    }
-    if !queue.runnable {
-        root.runnable_queues = root.runnable_queues.saturating_add(1);
-    }
-    mark_cos_queue_runnable(queue);
-}
-
-// #710: count an exact-drain TX submit stall on a specific queue.
-// NOT packet loss — on the exact path, `writer.insert == 0` leaves
-// the FIFO items in `queue.items` or restores them (flow-fair path);
-// frames that had been copied into UMEM are released back to
-// `free_tx_frames`, and the items get another chance next drain tick.
-// The counter signals TX-ring / completion-reap pressure, which is
-// an upstream cause for the downstream effects operators chase
-// (#706 mutex contention, #709 owner-worker hotspot).
-//
-// Non-exact transmit paths (`transmit_batch`, `transmit_prepared_queue`)
-// do not carry queue identity at the submit site and do not reach
-// this helper. Their frame-level failures are counted in the binding-
-// level `tx_submit_error_drops` counter instead.
-#[inline]
-pub(in crate::afxdp) fn count_tx_ring_full_submit_stall(
-    binding: &mut BindingWorker,
-    root_ifindex: i32,
-    queue_idx: usize,
-    stalled_packets: u64,
-) {
-    if stalled_packets == 0 {
-        return;
-    }
-    if let Some(root) = binding.cos_interfaces.get_mut(&root_ifindex) {
-        if let Some(queue) = root.queues.get_mut(queue_idx) {
-            queue.drop_counters.tx_ring_full_submit_stalls = queue
-                .drop_counters
-                .tx_ring_full_submit_stalls
-                .wrapping_add(stalled_packets);
-        }
-    }
-}
-
-fn rearm_cos_queue(root: &mut CoSInterfaceRuntime, queue_idx: usize, wake_tick: u64) {
-    park_cos_queue(root, queue_idx, wake_tick);
-}
-
-fn mark_cos_queue_runnable(queue: &mut CoSQueueRuntime) {
-    queue.runnable = true;
-    queue.parked = false;
-    queue.next_wakeup_tick = 0;
-}
-
-fn normalize_cos_queue_state(queue: &mut CoSQueueRuntime) {
-    if cos_queue_is_empty(queue) {
-        queue.runnable = false;
-        queue.parked = false;
-        queue.next_wakeup_tick = 0;
-        queue.surplus_deficit = 0;
-        return;
-    }
-    // Non-empty queues have only two valid steady states:
-    // 1. parked with a wakeup tick
-    // 2. runnable immediately
-    // Anything else can strand backlog forever.
-    if queue.parked && queue.next_wakeup_tick > 0 {
-        queue.runnable = false;
-        return;
-    }
-    mark_cos_queue_runnable(queue);
-}
-
-pub(in crate::afxdp) fn advance_cos_timer_wheel(root: &mut CoSInterfaceRuntime, now_ns: u64) {
-    let now_tick = cos_tick_for_ns(now_ns);
-    while root.timer_wheel.current_tick < now_tick {
-        root.timer_wheel.current_tick = root.timer_wheel.current_tick.saturating_add(1);
-        if root.timer_wheel.current_tick % COS_TIMER_WHEEL_L0_SLOTS as u64 == 0 {
-            cascade_cos_timer_wheel_level1(root);
-        }
-        wake_due_cos_timer_slot(root);
-    }
-}
-
-fn cascade_cos_timer_wheel_level1(root: &mut CoSInterfaceRuntime) {
-    let slot = ((root.timer_wheel.current_tick / COS_TIMER_WHEEL_L0_SLOTS as u64)
-        % COS_TIMER_WHEEL_L1_SLOTS as u64) as usize;
-    let queued = core::mem::take(&mut root.timer_wheel.level1[slot]);
-    let mut rearm = Vec::with_capacity(queued.len());
-    for queue_idx in queued {
-        let Some(queue) = root.queues.get(queue_idx) else {
-            continue;
-        };
-        if !queue.parked || queue.wheel_level != 1 || queue.wheel_slot != slot {
-            continue;
-        }
-        rearm.push((queue_idx, queue.next_wakeup_tick));
-    }
-    for (queue_idx, wake_tick) in rearm {
-        rearm_cos_queue(root, queue_idx, wake_tick);
-    }
-}
-
-fn wake_due_cos_timer_slot(root: &mut CoSInterfaceRuntime) {
-    let slot = (root.timer_wheel.current_tick % COS_TIMER_WHEEL_L0_SLOTS as u64) as usize;
-    let queued = core::mem::take(&mut root.timer_wheel.level0[slot]);
-    let mut rearm = Vec::with_capacity(queued.len());
-    let mut wake = Vec::with_capacity(queued.len());
-    for queue_idx in queued {
-        let Some(queue) = root.queues.get(queue_idx) else {
-            continue;
-        };
-        if !queue.parked || queue.wheel_level != 0 || queue.wheel_slot != slot {
-            continue;
-        }
-        if queue.next_wakeup_tick <= root.timer_wheel.current_tick {
-            wake.push(queue_idx);
-        } else {
-            rearm.push((queue_idx, queue.next_wakeup_tick));
-        }
-    }
-    for queue_idx in wake {
-        wake_cos_queue(root, queue_idx);
-    }
-    for (queue_idx, wake_tick) in rearm {
-        rearm_cos_queue(root, queue_idx, wake_tick);
-    }
-}
+// #956 P1: TX-completion + timer-wheel items reached by
+// `mod tests { use super::*; }` at the bottom of this file, plus
+// `redirect_local_cos_request_to_owner` which is also test-only after
+// the cross_binding extraction in #956 Phase 8.
+#[cfg(test)]
+use super::cos::{
+    advance_cos_timer_wheel, cos_queue_push_front, maybe_top_up_cos_root_lease,
+    normalize_cos_queue_state, park_cos_queue,
+    prepared_cos_request_stays_on_current_tx_binding, redirect_local_cos_request_to_owner,
+    restore_cos_local_items_inner, restore_cos_prepared_items_inner,
+    CoSServicePhase, COS_TIMER_WHEEL_TICK_NS,
+};
 
 pub(super) fn resolve_cos_queue_id(
     forwarding: &ForwardingState,
@@ -2111,206 +1882,6 @@ fn enqueue_cos_item(
     Ok(())
 }
 
-pub(in crate::afxdp) fn refresh_cos_interface_activity(binding: &mut BindingWorker, root_ifindex: i32) {
-    let mut new_nonempty = 0usize;
-    let mut new_runnable = 0usize;
-    let mut released_queue_leases = Vec::<(usize, u64)>::new();
-    let old_nonempty = binding
-        .cos_interfaces
-        .get(&root_ifindex)
-        .map(|root| root.nonempty_queues)
-        .unwrap_or(0);
-    if let Some(root) = binding.cos_interfaces.get_mut(&root_ifindex) {
-        for (queue_idx, queue) in root.queues.iter_mut().enumerate() {
-            normalize_cos_queue_state(queue);
-            if cos_queue_is_empty(queue) && queue.exact && queue.tokens > 0 {
-                released_queue_leases.push((queue_idx, core::mem::take(&mut queue.tokens)));
-            }
-            if cos_queue_is_empty(queue) {
-                continue;
-            }
-            new_nonempty = new_nonempty.saturating_add(1);
-            if queue.runnable {
-                new_runnable = new_runnable.saturating_add(1);
-            }
-        }
-        root.nonempty_queues = new_nonempty;
-        root.runnable_queues = new_runnable;
-    }
-    if old_nonempty == 0 && new_nonempty > 0 {
-        binding.cos_nonempty_interfaces = binding.cos_nonempty_interfaces.saturating_add(1);
-    } else if old_nonempty > 0 && new_nonempty == 0 {
-        binding.cos_nonempty_interfaces = binding.cos_nonempty_interfaces.saturating_sub(1);
-        release_cos_root_lease(binding, root_ifindex);
-    }
-    if let Some(iface_fast) = binding.cos_fast_interfaces.get(&root_ifindex) {
-        for (queue_idx, released) in released_queue_leases {
-            if let Some(shared_queue_lease) = iface_fast
-                .queue_fast_path
-                .get(queue_idx)
-                .and_then(|queue_fast| queue_fast.shared_queue_lease.as_ref())
-            {
-                shared_queue_lease.release_unused(released);
-            }
-        }
-    }
-}
-
-pub(in crate::afxdp) fn apply_cos_send_result(
-    binding: &mut BindingWorker,
-    root_ifindex: i32,
-    queue_idx: usize,
-    phase: CoSServicePhase,
-    batch_bytes: u64,
-    sent_bytes: u64,
-    retry: VecDeque<TxRequest>,
-) {
-    let mut exact_queue_idx = None;
-    {
-        let Some(root) = binding.cos_interfaces.get_mut(&root_ifindex) else {
-            return;
-        };
-        if let Some(queue) = root.queues.get_mut(queue_idx) {
-            exact_queue_idx = queue.exact.then_some(queue_idx);
-            let retry_bytes = restore_cos_local_items_inner(queue, retry);
-            queue.queued_bytes = queue
-                .queued_bytes
-                .saturating_sub(batch_bytes)
-                .saturating_add(retry_bytes);
-            match phase {
-                CoSServicePhase::Guarantee => {
-                    queue.tokens = queue.tokens.saturating_sub(sent_bytes);
-                }
-                CoSServicePhase::Surplus => {
-                    queue.surplus_deficit = queue.surplus_deficit.saturating_sub(sent_bytes);
-                }
-            }
-            // #760 instrumentation: record non-exact / surplus /
-            // shared-exact sends at the same site the queue's token
-            // or surplus accounting is debited. Paired with the
-            // apply_direct_exact_send_result write so the sum across
-            // all sites equals the bytes the CoS scheduler accounted.
-            queue
-                .owner_profile
-                .drain_sent_bytes
-                .fetch_add(sent_bytes, Ordering::Relaxed);
-        }
-        root.tokens = root.tokens.saturating_sub(sent_bytes);
-    }
-    if let Some(shared_root_lease) = binding
-        .cos_fast_interfaces
-        .get(&root_ifindex)
-        .and_then(|iface_fast| iface_fast.shared_root_lease.as_ref())
-    {
-        shared_root_lease.consume(sent_bytes);
-    }
-    if let Some(queue_idx) = exact_queue_idx {
-        if let Some(shared_queue_lease) = binding
-            .cos_fast_interfaces
-            .get(&root_ifindex)
-            .and_then(|iface_fast| iface_fast.queue_fast_path.get(queue_idx))
-            .and_then(|queue_fast| queue_fast.shared_queue_lease.as_ref())
-        {
-            shared_queue_lease.consume(sent_bytes);
-        }
-    }
-    refresh_cos_interface_activity(binding, root_ifindex);
-}
-
-pub(in crate::afxdp) fn apply_cos_prepared_result(
-    binding: &mut BindingWorker,
-    root_ifindex: i32,
-    queue_idx: usize,
-    phase: CoSServicePhase,
-    batch_bytes: u64,
-    sent_bytes: u64,
-    retry: VecDeque<PreparedTxRequest>,
-) {
-    let mut exact_queue_idx = None;
-    {
-        let Some(root) = binding.cos_interfaces.get_mut(&root_ifindex) else {
-            return;
-        };
-        if let Some(queue) = root.queues.get_mut(queue_idx) {
-            exact_queue_idx = queue.exact.then_some(queue_idx);
-            let retry_bytes = restore_cos_prepared_items_inner(queue, retry);
-            queue.queued_bytes = queue
-                .queued_bytes
-                .saturating_sub(batch_bytes)
-                .saturating_add(retry_bytes);
-            match phase {
-                CoSServicePhase::Guarantee => {
-                    queue.tokens = queue.tokens.saturating_sub(sent_bytes);
-                }
-                CoSServicePhase::Surplus => {
-                    queue.surplus_deficit = queue.surplus_deficit.saturating_sub(sent_bytes);
-                }
-            }
-            // #760 instrumentation, the FOURTH apply_* site. This is
-            // the prepared-batch path (CoSBatch::Prepared, in-place
-            // rewrite — the common case for forwarded traffic). The
-            // initial instrumentation commit missed this site; the
-            // first 120 s iperf3 measurement showed only ~987 Mbps
-            // on drain_sent_bytes while the receiver reported 1.55
-            // Gbps, leaving ~563 Mbps unaccounted — all of it
-            // flowing through this path. Same Relaxed semantics as
-            // the other three apply_* sites.
-            queue
-                .owner_profile
-                .drain_sent_bytes
-                .fetch_add(sent_bytes, Ordering::Relaxed);
-        }
-        root.tokens = root.tokens.saturating_sub(sent_bytes);
-    }
-    if let Some(shared_root_lease) = binding
-        .cos_fast_interfaces
-        .get(&root_ifindex)
-        .and_then(|iface_fast| iface_fast.shared_root_lease.as_ref())
-    {
-        shared_root_lease.consume(sent_bytes);
-    }
-    if let Some(queue_idx) = exact_queue_idx {
-        if let Some(shared_queue_lease) = binding
-            .cos_fast_interfaces
-            .get(&root_ifindex)
-            .and_then(|iface_fast| iface_fast.queue_fast_path.get(queue_idx))
-            .and_then(|queue_fast| queue_fast.shared_queue_lease.as_ref())
-        {
-            shared_queue_lease.consume(sent_bytes);
-        }
-    }
-    refresh_cos_interface_activity(binding, root_ifindex);
-}
-
-pub(in crate::afxdp) fn restore_cos_local_items_inner(
-    queue: &mut CoSQueueRuntime,
-    mut retry: VecDeque<TxRequest>,
-) -> u64 {
-    let mut retry_bytes = 0u64;
-    while let Some(req) = retry.pop_back() {
-        retry_bytes = retry_bytes.saturating_add(req.bytes.len() as u64);
-        cos_queue_push_front(queue, CoSPendingTxItem::Local(req));
-    }
-    if !cos_queue_is_empty(queue) {
-        mark_cos_queue_runnable(queue);
-    }
-    retry_bytes
-}
-
-pub(in crate::afxdp) fn restore_cos_prepared_items_inner(
-    queue: &mut CoSQueueRuntime,
-    mut retry: VecDeque<PreparedTxRequest>,
-) -> u64 {
-    let mut retry_bytes = 0u64;
-    while let Some(req) = retry.pop_back() {
-        retry_bytes = retry_bytes.saturating_add(req.len as u64);
-        cos_queue_push_front(queue, CoSPendingTxItem::Prepared(req));
-    }
-    if !cos_queue_is_empty(queue) {
-        mark_cos_queue_runnable(queue);
-    }
-    retry_bytes
-}
 
 fn process_pending_queue_in_place<T, F>(pending: &mut VecDeque<T>, mut f: F)
 where
